separate file directory into import and export directories
