+++ console/src/export/index.ts
@@ -7,4 +7,4 @@
 // Source License, use of this software will be governed by the Apache License,
 // Version 2.0, included in the file licenses/APL.txt.
 
-export * as File from "@/file/external";
+export * as Export from "@/export/external";
