+++ console/src/extractors.ts
@@ -7,13 +7,13 @@
 // License, use of this software will be governed by the Apache License, Version 2.0,
 // included in the file licenses/APL.txt.
 
-import { type File } from "@/file";
+import { type Export } from "@/export";
 import { LinePlot } from "@/lineplot";
 import { Log } from "@/log";
 import { Schematic } from "@/schematic";
 import { Table } from "@/table";
 
-export const EXTRACTORS: Record<string, File.Extractor> = {
+export const EXTRACTORS: Record<string, Export.Extractor> = {
   [LinePlot.LAYOUT_TYPE]: LinePlot.extract,
   [Log.LAYOUT_TYPE]: Log.extract,
   [Schematic.LAYOUT_TYPE]: Schematic.extract,
