+++ console/src/lineplot/services/import.ts
@@ -16,7 +16,7 @@ import { readFile } from "@tauri-apps/plugin-fs";
 import { useDispatch, useStore } from "react-redux";
 
 import { Confirm } from "@/confirm";
-import { type File } from "@/file";
+import { type Import } from "@/import";
 import { Layout } from "@/layout";
 import { create } from "@/lineplot/LinePlot";
 import { select } from "@/lineplot/selectors";
@@ -24,7 +24,7 @@ import { parser, remove } from "@/lineplot/slice";
 import { type RootState } from "@/store";
 import { Workspace } from "@/workspace";
 
-export const fileHandler: File.FileHandler = async ({
+export const fileHandler: Import.FileHandler = async ({
   file,
   place,
   tab,
@@ -86,7 +86,7 @@ export const fileHandler: File.FileHandler = async ({
 
 const filters: DialogFilter[] = [{ name: "JSON", extensions: ["json"] }];
 
-interface ImportPlotProps extends Omit<File.FileHandlerProps, "file" | "name"> {
+interface ImportPlotProps extends Omit<Import.FileHandlerProps, "file" | "name"> {
   workspaceKey?: string;
   activeWorkspaceKey: string | null;
 }
