+++ console/src/lineplot/toolbar/Toolbar.tsx
@@ -17,7 +17,7 @@ import { useDispatch } from "react-redux";
 
 import { ToolbarHeader, ToolbarTitle } from "@/components";
 import { CSS } from "@/css";
-import { File } from "@/file";
+import { Export } from "@/export";
 import { Layout } from "@/layout";
 import { useExport } from "@/lineplot/export";
 import { useSelect, useSelectToolbar } from "@/lineplot/selectors";
@@ -90,7 +90,7 @@ export const Toolbar = ({ layoutKey }: ToolbarProps): ReactElement | null => {
           <ToolbarTitle icon={<Icon.Visualize />}>{name}</ToolbarTitle>
           <Align.Space direction="x" align="center" empty>
             <Align.Space direction="x" empty style={{ height: "100%", width: 66 }}>
-              <File.ExportToolbarButton onExport={() => handleExport(state.key)} />
+              <Export.ToolbarButton onExport={() => handleExport(state.key)} />
               <Link.ToolbarCopyButton
                 name={name}
                 ontologyID={linePlot.ontologyID(state.key)}
